Move start stop functions into nginx scripts to ensure nginx scripts can run as standalones

+++ scripts/nginx/deluge.sh
@@ -24,11 +24,18 @@ fi
 for u in "${users[@]}"; do
   if [[ -f /etc/nginx/apps/${u}.dindex.conf ]]; then rm -f /etc/nginx/apps/${u}.dindex.conf; fi
 
-systemctl stop deluge-web@$u
-sed -i 's/"interface": "0.0.0.0"/"interface": "127.0.0.1"/g' /home/$u/.config/deluge/web.conf
-sed -i 's/"https": true/"https": false/g' /home/$u/.config/deluge/web.conf
-systemctl start deluge-web@$u
+  isactive=$(systemctl is-active deluge-web@$u)
+  if [[ $isactive == "active" ]]; then
+    systemctl stop deluge-web@$u
+  fi
 
+  sed -i 's/"interface": "0.0.0.0"/"interface": "127.0.0.1"/g' /home/$u/.config/deluge/web.conf
+  sed -i 's/"https": true/"https": false/g' /home/$u/.config/deluge/web.conf
+
+  if [[ $isactive == "active" ]]; then
+    systemctl start deluge-web@$u
+  fi
+  
   if [[ ! -f /etc/nginx/conf.d/${u}.deluge.conf ]]; then
     DWPORT=$(cat /home/$u/.config/deluge/web.conf | grep port | cut -d: -f2| sed 's/ //g' | sed 's/,//g')
     cat > /etc/nginx/conf.d/${u}.deluge.conf <<DUPS
